web/pyproject.toml
[tool.poetry]
name = "caterchain_poc"
version = "0.1.0"
description = ""
authors = ["Brian Ibbotson <brian@quinceleaf.dev>"]

[tool.poetry.dependencies]
python = "^3.8"
Django = "^3.2"
psycopg2-binary = "^2.8.6"
gunicorn = "^20.1.0"
django-redis = "^4.12.1"
django-admin-autocomplete-filter = "^0.6.1"
django-filter = "^2.4.0"
django-formtools = "^2.2"
django-fsm = "^2.7.1"
django-fsm-log = {git = "https://github.com/quinceleaf/django-fsm-log.git"}
django-import-export = "^2.5.0"
django-pandas = "^0.6.4"
django-select2 = "^7.6.2"
django-simple-history = "^2.12.0"
django-treebeard = "^4.5.1"
django-widget-tweaks = "^1.4.8"
djangorestframework = "^3.12.2"
numpy = "^1.20.2"
openpyxl = "^3.0.7"
pandas = "^1.2.3"
Pillow = "^8.1.2"
python-dotenv = "^0.15.0"
python-ulid = "^1.0.2"
pytz = "^2021.1"
celery = "5.0.5"
django-debug-toolbar = "^3.2"
coverage = "^5.5"
django-celery-beat = "^2.2.0"
aloe-django = "^0.2.0"
django-coverage-plugin = "^1.8.0"
weasyprint = "^52.5"



[tool.poetry.dev-dependencies]
black = "^20.8b1"
Faker = "^6.6.0"
factory-boy = "^3.2.0"
flake8 = "^3.8.4"
model-bakery = "^1.2.1"
selenium = "^3.141.0"
PyNaCl = "^1.4.0"
requests = "^2.25.1"


[build-system]
requires = ["poetry-core==1.0.0","setuptools>=50", "cryptography<=3.3.2"]
build-backend = "poetry.core.masonry.api"
